Move one more class
